refactor: move preset helpers to backend module
