+++ backend/__init__.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+DEFAULT_SETS_PER_EXERCISE = 3
+DEFAULT_REST_DURATION = 120
+DEFAULT_DB_PATH = Path(__file__).resolve().parent.parent / "data" / "workout.db"
