+++ backend/workout_session.py
@@ -2,13 +2,12 @@ import sqlite3
 import time
 from pathlib import Path
 
-from core import (
-    get_metrics_for_exercise,
-    get_metrics_for_preset,
+from backend import (
     DEFAULT_SETS_PER_EXERCISE,
     DEFAULT_REST_DURATION,
     DEFAULT_DB_PATH,
 )
+from backend.metrics import get_metrics_for_exercise, get_metrics_for_preset
 
 
 class WorkoutSession:
