+++ tests/test_enum_values.py
@@ -1,7 +1,6 @@
 import sqlite3
 from pathlib import Path
-import core
-from backend import exercises
+from backend import exercises, metrics
 from backend.exercise import Exercise
 
 
@@ -14,13 +13,13 @@ def test_enum_values_default_and_override(sample_db: Path) -> None:
     conn.close()
 
     # Add Machine metric to an exercise without override
-    core.add_metric_to_exercise("Push-up", "Machine", db_path=sample_db)
+    metrics.add_metric_to_exercise("Push-up", "Machine", db_path=sample_db)
 
-    metrics_push = core.get_metrics_for_exercise("Push-up", db_path=sample_db)
+    metrics_push = metrics.get_metrics_for_exercise("Push-up", db_path=sample_db)
     vals_push = next(m["values"] for m in metrics_push if m["name"] == "Machine")
     assert vals_push == ["A", "B", "C"]
 
-    metrics_bench = core.get_metrics_for_exercise("Bench Press", db_path=sample_db)
+    metrics_bench = metrics.get_metrics_for_exercise("Bench Press", db_path=sample_db)
     vals_bench = next(m["values"] for m in metrics_bench if m["name"] == "Machine")
     assert vals_bench == ["A", "B"]
 
@@ -29,6 +28,6 @@ def test_enum_values_default_and_override(sample_db: Path) -> None:
     ex.update_metric("Machine", values=["A"])
     exercises.save_exercise(ex)
 
-    metrics_bench2 = core.get_metrics_for_exercise("Bench Press", db_path=sample_db)
+    metrics_bench2 = metrics.get_metrics_for_exercise("Bench Press", db_path=sample_db)
     vals_bench2 = next(m["values"] for m in metrics_bench2 if m["name"] == "Machine")
     assert vals_bench2 == ["A"]
