+++ ui/screens/edit_preset_screen.py
@@ -26,9 +26,8 @@ from kivymd.uix.dialog import MDDialog
 import os
 import sqlite3
 
-import core
-from backend import exercises
-from core import DEFAULT_DB_PATH
+from backend import DEFAULT_DB_PATH, DEFAULT_SETS_PER_EXERCISE
+from backend import exercises, metrics, presets
 
 
 class SectionWidget(MDBoxLayout):
@@ -327,7 +326,7 @@ class EditPresetScreen(MDScreen):
                 else:
                     results = []
                     description = ""
-                    metric_defs = core.get_metrics_for_exercise(
+                    metric_defs = metrics.get_metrics_for_exercise(
                         ex["name"],
                         db_path=session.db_path,
                         preset_name=session.preset_name,
@@ -335,7 +334,7 @@ class EditPresetScreen(MDScreen):
                 new_exercises.append(
                     {
                         "name": ex.get("name"),
-                        "sets": ex.get("sets") or core.DEFAULT_SETS_PER_EXERCISE,
+                        "sets": ex.get("sets") or DEFAULT_SETS_PER_EXERCISE,
                         "rest": ex.get("rest") or session.rest_duration,
                         "results": results,
                         "library_exercise_id": ex.get("library_id"),
@@ -562,7 +561,7 @@ class EditPresetScreen(MDScreen):
 
         all_defs = {
             m["name"]: m
-            for m in core.get_all_metric_types(include_user_created=True)
+            for m in metrics.get_all_metric_types(include_user_created=True)
         }
 
         rv.data = [
@@ -613,7 +612,7 @@ class EditPresetScreen(MDScreen):
         def do_confirm(*args):
             try:
                 app.preset_editor.save()
-                core.load_workout_presets(app.preset_editor.db_path)
+                presets.load_workout_presets(app.preset_editor.db_path)
                 app.selected_preset = app.preset_editor.preset_name
                 if dialog:
                     dialog.dismiss()
@@ -913,7 +912,7 @@ class AddPresetMetricPopup(MDDialog):
             }
         metrics = [
             m
-            for m in core.get_all_metric_types()
+            for m in metrics.get_all_metric_types()
             if m.get("scope") == "preset" and (
                 m.get("name") not in existing and m.get("metric_name") not in existing
             )
@@ -965,7 +964,7 @@ class AddSessionMetricPopup(MDDialog):
             }
         metrics = [
             m
-            for m in core.get_all_metric_types()
+            for m in metrics.get_all_metric_types()
             if m.get("scope") == "session" and (
                 m.get("name") not in existing and m.get("metric_name") not in existing
             )
